added confirmatoin messages
to modal dialogs

+++ src/main/java/org/strasa/web/admin/view/model/EditStudyVariable.java
@@ -28,6 +28,8 @@ import org.zkoss.bind.annotation.Init;
 import org.zkoss.bind.annotation.NotifyChange;
 import org.zkoss.zk.ui.Component;
 import org.zkoss.zk.ui.Executions;
+import org.zkoss.zk.ui.event.Event;
+import org.zkoss.zk.ui.event.EventListener;
 import org.zkoss.zul.Div;
 import org.zkoss.zul.Messagebox;
 import org.zkoss.zul.Tabpanel;
@@ -87,13 +89,25 @@ public class EditStudyVariable {
 		BindUtils.postNotifyChange(null, null, ps, "editingStatus");
 	}
 
+	@SuppressWarnings("unchecked")
 	@NotifyChange("rowList")
 	@Command("delete")
-	public void delete(@BindingParam("id") Integer Id, @BindingParam("varCode") String varCode){
+	public void delete(@BindingParam("id") final Integer Id, @BindingParam("varCode") String varCode){
 		if(!studyDataColMan.existsStudyDataColumnByName(varCode)){
-		man.deleteById(Id);
-		makeRowStatus(man.getVariables());
-		Messagebox.show("Changes saved.");
+		Messagebox.show("Are you sure?",
+				"Delete", Messagebox.OK | Messagebox.CANCEL,
+				Messagebox.QUESTION, new EventListener() {
+			public void onEvent(Event e) {
+				if ("onOK".equals(e.getName())) {
+					man.deleteById(Id);
+					makeRowStatus(man.getVariables());
+					BindUtils.postNotifyChange(null, null,
+							EditStudyVariable.this, "rowList");
+					Messagebox.show("Changes saved.");
+				} else if ("onCancel".equals(e.getName())) {
+				}
+			}
+		});
 		} else  Messagebox.show("Cannot delete a study variable that is in use.", "Error", Messagebox.OK, Messagebox.ERROR); 
 	}
